% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.r
\name{formatAccept}
\alias{formatAccept}
\title{Format the data for the 'Accept' header}
\usage{
formatAccept(version, contentType)
}
\arguments{
\item{version}{version}

\item{contentType}{contentType}
}
\description{
Format the data for the 'Accept' header
}
\examples{
formatAccept("2.0", "text/xml")
}

