% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg.r
\name{getCfg}
\alias{getCfg}
\title{Get the LCMAP client configuration as a hashmap}
\usage{
getCfg(filename)
}
\arguments{
\item{filename}{filename}
}
\description{
Get the LCMAP client configuration as a hashmap
}
\examples{
getCfg()
getCfg("~/.usgs/lcmap.ini")
}
\seealso{
Other config: \code{\link{getIni}}
}

