% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmap.r
\docType{package}
\name{lcmap}
\alias{clientVersion}
\alias{lcmap}
\alias{lcmap-package}
\title{\pkg{lcmap} R client for LCMAP REST services}
\format{An object of class \code{character} of length 1.}
\usage{
clientVersion
}
\description{
\pkg{lcmap} R client for LCMAP REST services
}
\keyword{datasets}

