% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{login}
\alias{login}
\title{Authenticate to LCMAP using ERS}
\usage{
login(username, password, version)
}
\arguments{
\item{username}{username}

\item{password}{password}

\item{version}{version}
}
\description{
Authenticate to LCMAP using ERS
}
\examples{
login("alice", "secret")
login("alice", "secret", "2.0")
}

