% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfg.r
\name{getIni}
\alias{getIni}
\title{Read the LCMAP client configuration file}
\usage{
getIni(filename)
}
\arguments{
\item{filename}{filename}

\item{password}{password}
}
\description{
Read the LCMAP client configuration file
}
\examples{
getIni()
getIni("custom/loc/lcmap.ini")
}
\seealso{
Other config: \code{\link{getCfg}}
}

